third_party/WebKit/LayoutTests/platform/win/http/tests/security/powerfulFeatureRestrictions/geolocation-on-secure-origin-in-secure-origin-expected.txt

CONSOLE ERROR: line 11: This test can not run without mojo
CONSOLE ERROR: line 11: This test can not run without mojo
CONSOLE ERROR: line 11: This test can not run without mojo
CONSOLE ERROR: line 11: This test can not run without mojo
CONSOLE ERROR: line 11: This test can not run without mojo
This is a testharness.js-based test.
PASS getCurrentPosition in iframe 
Harness: the test ran to completion.
